Move HOC to subfolder
