+++ assets/src/components/higher-order/with-video-poster-image-notice.js
